Fix document type codes per Hacienda ANEXOS v4.4
Corrected NC from 02 to 03, ND from 03 to 02, and REP from 11 to 10
per official ANEXOS v4.4 TipoDocReferenciaType and Digifact documentation.
NC=03 (Nota de Crédito), ND=02 (Nota de Débito), REP=10 (Recibo Electrónico de Pago).

+++ mercurius-quarkus/src/main/java/Controllers/Tiquetes/CrearTiqueteController.java
@@ -245,7 +245,7 @@ public class CrearTiqueteController implements Serializable {
         carritoService.setColones(total);
         carritoService.setDolares(BigDecimal.ZERO);
         carritoService.setTotalPagado(total);
-        carritoService.calcularVuelto(BigDecimal.valueOf(tipoCambio.getTipoCambioActual().getValorCompra()));
+        carritoService.calcularVuelto(tipoCambio.getTipoCambioActual().getValorCompra());
     }
 
     @Nonnull
